[BPK-4298] Code cleanup
