+++ Example/Backpack/BPKAppDelegate.m
@@ -37,25 +37,6 @@
 @implementation BPKAppDelegate
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
-    self.window = [[UIWindow alloc] initWithFrame:UIScreen.mainScreen.bounds];
-    UINavigationController *navigationController = [UINavigationController new];
-
-    BPKRootTableViewController *rootTableViewController = [BPKRootTableViewController new];
-    navigationController.viewControllers = @[rootTableViewController];
-    self.window.rootViewController = navigationController;
-
-    if ([ThemeHelpers isThemingSupported]) {
-        [ThemeHelpers applyAllThemes];
-
-        id<BPKThemeDefinition> activeTheme = [ThemeHelpers themeDefinitionForTheme:Settings.sharedSettings.activeTheme];
-
-        BPKThemeContainerController *themeContainerController =
-            [[BPKThemeContainerController alloc] initWithThemeDefinition:activeTheme rootViewController:self.window.rootViewController];
-        self.window.rootViewController = themeContainerController;
-    }
-
-    [self.window makeKeyAndVisible];
-
     self.UITestingEnabled = [NSProcessInfo.processInfo.arguments containsObject:@"UITests"];
 
     id<BPKFontDefinitionProtocol> relativeFontDefinition = [BPKRelativeFontDefinition new];
@@ -83,6 +64,25 @@
     };
 
     [BPKAppearance apply];
+
+    self.window = [[UIWindow alloc] initWithFrame:UIScreen.mainScreen.bounds];
+    UINavigationController *navigationController = [UINavigationController new];
+
+    BPKRootTableViewController *rootTableViewController = [BPKRootTableViewController new];
+    navigationController.viewControllers = @[rootTableViewController];
+    self.window.rootViewController = navigationController;
+
+    if ([ThemeHelpers isThemingSupported]) {
+        [ThemeHelpers applyAllThemes];
+
+        id<BPKThemeDefinition> activeTheme = [ThemeHelpers themeDefinitionForTheme:Settings.sharedSettings.activeTheme];
+
+        BPKThemeContainerController *themeContainerController =
+            [[BPKThemeContainerController alloc] initWithThemeDefinition:activeTheme rootViewController:self.window.rootViewController];
+        self.window.rootViewController = themeContainerController;
+    }
+
+    [self.window makeKeyAndVisible];
 
     // If simctl gets the ability to handle this then we should switch to using it.
     if ([ThemeHelpers forceDarkMode]) {
