+++ Example/Backpack/NavigationData.swift
@@ -16,155 +16,8 @@
  * limitations under the License.
  */
 
-protocol Presentable {
-    func makeViewController() -> UIViewController
-}
-
-enum ItemValue {
-    case story(Presentable)
-    case group([Item])
-}
-
-protocol ItemConvertible {
-    func asItems() -> [Item]
-}
-
-public struct Item {
-    var name: String
-    var value: ItemValue
-}
-
-extension Item {
-    func isGroup() -> Bool {
-        switch self.value {
-        case .group:
-            return true
-        default:
-            return false
-        }
-    }
-
-    func isStory() -> Bool {
-        switch self.value {
-        case .story:
-            return true
-        default:
-            return false
-        }
-    }
-
-    func subItems() -> [Item]? {
-        switch self.value {
-        case .group(let items):
-            return items
-        default:
-            return nil
-        }
-    }
-}
-
-extension Item: ItemConvertible {
-    func asItems() -> [Item] {
-        return [self]
-    }
-}
-
-public struct Group {
-    var name: String
-    var items: [Item]
-
-    init(name: String, @AppBuilder builder: () -> [Item]) {
-        self.name = name
-        self.items = builder()
-    }
-}
-
-extension Group: ItemConvertible {
-    func asItems() -> [Item] {
-        return [Item(name: name, value: .group(items))]
-    }
-}
-
-extension Array: ItemConvertible where Element == Item {
-    func asItems() -> [Item] {
-        self
-    }
-}
-
-@_functionBuilder
-public struct AppBuilder {
-    // returns empty array of Items
-    public static func buildBlock() -> [Item] { [] }
-}
-
-extension AppBuilder {
-    // maps groups to list of Items
-    static func buildBlock(_ groups: ItemConvertible...) -> [Item] {
-        groups.flatMap { $0.asItems() }
-    }
-}
-
-public func makeApp(@AppBuilder _ content: () -> [Item]) -> [Item] {
-    content()
-}
-
-struct StoryboardPresentable: Presentable {
-    func makeViewController() -> UIViewController {
-        let viewController: UIViewController =
-            storyboard.instantiateViewController(withIdentifier: identifier)
-        viewController.modalPresentationStyle = UIModalPresentationStyle.fullScreen
-        return viewController
-    }
-
-    var storyboard: UIStoryboard
-    var identifier: String
-}
-
-func loadStoryboard(name: String) -> (String) -> Presentable {
-    let storyboard = UIStoryboard(name: name, bundle: nil)
-
-    return { viewControllerIdentier in
-         return StoryboardPresentable(storyboard: storyboard, identifier: viewControllerIdentier)
-    }
-}
-
-func loadStoryboard(name: String, identifier: String) -> Presentable {
-    return loadStoryboard(name: name)(identifier)
-}
-
-public struct Section<T> {
-    let name: String?
-    let rows: [Row<T>]
-}
-
-public struct Row<T> {
-    let name: String
-    let value: T
-}
-
-public func sectionify(items: [Item]) -> [Section<Item>] {
-    let groups: [(String, [Item])] = items.compactMap({ group in
-            group.subItems().map({
-                (group.name, $0)
-            })
-    })
-
-    let ungroupedItems: [Item] = items.filter({ $0.isStory() })
-
-    var value = groups.map({group in
-        Section(name: group.0, rows: group.1.map({ Row(name: $0.name, value: $0) }))
-    })
-
-    if ungroupedItems.count > 0 {
-        value.append(Section(name: nil, rows: ungroupedItems.map({ Row(name: $0.name, value: $0) })))
-    }
-
-    return value
-}
-
 @objc
 class NavigationData: NSObject {
-
     static var mainStoryboard = loadStoryboard(name: "Main")
 
     // swiftlint:disable line_length closure_body_length
@@ -172,18 +25,7 @@ class NavigationData: NSObject {
     static var appStructure: [Item] = makeApp {
         Group(name: "Tokens") {
             Item(name: "Colors", value: .story(mainStoryboard("ColorsViewController")))
-            Group(name: "Gradients") {
-                Item(name: "Primary", value: .story(loadStoryboard(name: "Gradients", identifier: "GradientsViewController")))
-                Item(name: "Baseline scrim", value: .story(loadStoryboard(name: "Gradients", identifier: "GradientsViewController")))
-            }
-            Group(name: "Gradients nested") {
-                Group(name: "Nest") {
-                    Item(name: "Primary", value: .story(loadStoryboard(name: "Gradients", identifier: "GradientsViewController")))
-                }
-                Group(name: "Nest") {
-                    Item(name: "Baseline scrim", value: .story(loadStoryboard(name: "Gradients", identifier: "GradientsViewController")))
-                }
-            }
+            Item(name: "Gradients", value: .story(loadStoryboard(name: "Gradients", identifier: "GradientsViewController")))
             Item(name: "Spacings", value: .story(mainStoryboard("SpacingsViewController")))
             Item(name: "Radii", value: .story(mainStoryboard("RadiiViewController")))
             Item(name: "Shadows", value: .story(mainStoryboard("ShadowsViewController")))
@@ -195,8 +37,8 @@ class NavigationData: NSObject {
             Item(name: "Buttons", value: .story(loadStoryboard(name: "Buttons", identifier: "ButtonsViewController")))
             Item(name: "Calendar", value: .story(loadStoryboard(name: "Calendar", identifier: "CalendarViewController")))
             Item(name: "Cards", value: .story(loadStoryboard(name: "Cards", identifier: "CardsViewController")))
-            Item(name: "Chips", value: .story(loadStoryboard(name: "Chips", identifier: "ChipsViewController")))
             Item(name: "Flare views", value: .story(loadStoryboard(name: "FlareView", identifier: "FlareViewViewController")))
+            Item(name: "Chips", value: .story(loadStoryboard(name: "Chips", identifier: "ChipsViewController")))
             Item(name: "Dialogs", value: .story(loadStoryboard(name: "Dialogs", identifier: "DialogsViewController")))
             Item(name: "Horizontal navigation", value: .story(loadStoryboard(name: "HorizontalNavigation", identifier: "HorizontalNavigationViewController")))
             Item(name: "Icons", value: .story(mainStoryboard("IconsViewController")))
@@ -217,4 +59,7 @@ class NavigationData: NSObject {
             Item(name: "Toasts", value: .story(loadStoryboard(name: "Toasts", identifier: "ToastsViewController")))
         }
     }
+
+    // swiftlint:enable
+
 }
