+++ Example/Backpack/StoryboardPresentable.swift
@@ -1,0 +1,42 @@
+//
+/*
+ * Backpack - Skyscanner's Design System
+ *
+ * Copyright © 2021 Skyscanner Ltd. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+struct StoryboardPresentable: Presentable {
+    func makeViewController() -> UIViewController {
+        let viewController: UIViewController =
+            storyboard.instantiateViewController(withIdentifier: identifier)
+        viewController.modalPresentationStyle = UIModalPresentationStyle.fullScreen
+        return viewController
+    }
+
+    var storyboard: UIStoryboard
+    var identifier: String
+}
+
+func loadStoryboard(name: String) -> (String) -> Presentable {
+    let storyboard = UIStoryboard(name: name, bundle: nil)
+
+    return { viewControllerIdentier in
+         StoryboardPresentable(storyboard: storyboard, identifier: viewControllerIdentier)
+    }
+}
+
+func loadStoryboard(name: String, identifier: String) -> Presentable {
+    loadStoryboard(name: name)(identifier)
+}
